backend/requirements/base.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
channels==4.0.0
channels-redis==4.1.0
drf-spectacular==0.27.0  # API documentation
openai==1.7.0
python-telegram-bot==20.7
Pillow==10.1.0
requests==2.31.0
pandas==2.1.4
python-decouple==3.8
dj-database-url==2.1.0
whitenoise==6.6.0  # Static files
gunicorn==21.2.0